functions/detect_label/requirements.txt
google-cloud-storage==1.13.0
google-cloud-vision==0.39.0
google-cloud-firestore==1.5.0